backend/requirements.txt
flask
python-dotenv
flask-restful
psycopg2
PyJWT